Update after resync
